update index.html and styles.css
I finalized the html layout of the main component, may need some tweaking as I add more styling to the elements. I started to stylize the main component and had some issues with it aligning to the left instead of centering both vertically and horizontally.  Added my variables for the colors in the :root, will need to set up media queries to keep the card and image from breaking in mobile/desktop. May include a landscape orientation if I have enough time to.