updated styles.css
Tweaked the padding and margin issues, made it look as close to the challenge prompt as possible. Still need to add the background per a media query. Active states work as they should, I added a box (drop) shadow below the payment link and fixed the main container height to max-content. Next phase is to see what it looks like on mobile and adjust.